Go: fix alignment
