+++ internal/entity/llm_factories.go
@@ -16,18 +16,17 @@
 
 package entity
 
-// UserTenant user tenant relationship model
-type UserTenant struct {
-	ID        string  `gorm:"column:id;primaryKey;size:32" json:"id"`
-	UserID    string  `gorm:"column:user_id;size:32;not null;index" json:"user_id"`
-	TenantID  string  `gorm:"column:tenant_id;size:32;not null;index" json:"tenant_id"`
-	Role      string  `gorm:"column:role;size:32;not null;index" json:"role"`
-	InvitedBy string  `gorm:"column:invited_by;size:32;not null;index" json:"invited_by"`
-	Status    *string `gorm:"column:status;size:1;index;default:'1'" json:"status,omitempty"`
+// LLMFactories LLM factory model
+type LLMFactories struct {
+	Name   string  `gorm:"column:name;primaryKey;size:128" json:"name"`
+	Logo   *string `gorm:"column:logo;type:longtext" json:"logo,omitempty"`
+	Tags   string  `gorm:"column:tags;size:255;not null;index" json:"tags"`
+	Rank   int64   `gorm:"column:rank;default:0" json:"rank"`
+	Status *string `gorm:"column:status;size:1;index;default:'1'" json:"status,omitempty"`
 	BaseModel
 }
 
 // TableName specify table name
-func (UserTenant) TableName() string {
-	return "user_tenant"
+func (LLMFactories) TableName() string {
+	return "llm_factories"
 }
